Add seconds unit support to logger timing display + refactor

+++ srcs/utils/Logger.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Logger.hpp                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: eschwart <eschwart@student.42.fr>          +#+  +:+       +#+        */
+/*   By: gdosch <gdosch@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/09 11:33:46 by eschwart          #+#    #+#             */
-/*   Updated: 2026/02/12 10:31:35 by eschwart         ###   ########.fr       */
+/*   Updated: 2026/02/12 11:05:17 by gdosch           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,21 +29,22 @@
 
 // Class ***********************************************************************
 class Logger {
-	private:
-		// Attribute(s) --------------------------------------------------------
-		// static timeval _lastRequestTime;
-		static std::string _lastMethod; ///< Last logged HTTP method
-		static std::string _lastUri; ///< Last logged URI
-		static std::string _lastClientIP; ///< Last logged client IP address
-		static int _lastStatus; ///< Last logged HTTP status code
-		static size_t _lastSize; ///< Last logged response size
-		static int _requestCount; ///< Number of grouped identical requests
-		static double _totalTime; ///< Total response time for grouped requests
-		static double _minTime; ///< Minimum response time in group
-		static double _maxTime; ///< Maximum response time in group
-		static std::string _lastServerName; ///< Last logged server name
-		static int _lastServerPort; ///< Last logged server port
-		static bool _firstLog; ///< Indicates if this is the first log entry
+private:
+	// Attribute(s) ------------------------------------------------------------
+	// static timeval _lastRequestTime;
+	static std::string _lastMethod; ///< Last logged HTTP method
+	static std::string _lastUri; ///< Last logged URI
+	static std::string _lastClientIP; ///< Last logged client IP address
+	static int _lastStatus; ///< Last logged HTTP status code
+	static size_t _lastSize; ///< Last logged response size
+	static size_t _requestCount; ///< Number of grouped identical requests
+	static double _totalTime; ///< Total response time for grouped requests
+	static double _minTime; ///< Minimum response time in group
+	static double _maxTime; ///< Maximum response time in group
+	static std::string _lastServerName; ///< Last logged server name
+	static int _lastServerPort; ///< Last logged server port
+	static bool _firstLog; ///< Indicates if this is the first log entry
+
 
 		// Private method(s) ---------------------------------------------------
 		static std::string getCurrentTime();
